+++ R/cauphylm.R
@@ -345,7 +345,10 @@ AIC.cauphylm <- phylolm::AIC.phylolm
 #' @inheritParams phylolm::predict.phylolm
 #' @method predict cauphylm
 #' @rdname vcov.cauphylm
-predict.cauphylm <- phylolm::predict.phylolm
+predict.cauphylm <- function(object, newdata=NULL, se.fit = FALSE, ...){
+  if (se.fit) stop("'se.fit' cannot be used in 'predict.cauphylm'. Please set to 'se.fit=FALSE'.")
+    return(phylolm::predict.phylolm(object, newdata, se.fit = FALSE, ...))
+}
 #' @export
 #' @inheritParams stats::confint
 #' @method confint cauphylm