+++ NEWS.md
@@ -1,5 +1,7 @@
 # cauphy 1.0.2.9000
 
+* Make 'predict.cauphylm' compatible with 'predict.phylolm' version 2.6.5.
+
 # cauphy 1.0.2
 
 * Fix CRAN check results issues with r-devel.